revrite migrations, handle user-role depedency

+++ src/main/resources/db/migration/V7__create.user.role.relations.sql
@@ -13,7 +13,7 @@ INSERT INTO roles (name, user_system_name, created_by) VALUES
 ;
 
 CREATE Table users_roles (
-    user_id BIGSERIAL NOT NULL,
+    user_id INT NOT NULL,
     role_id INT NOT NULL,
     PRIMARY KEY (user_id, role_id),
     FOREIGN KEY (user_id) REFERENCES users (id),
